# services/orchestrator-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: MentatLab Orchestrator API
  description: |
    RESTful API for the MentatLab orchestrator service.

    ## Overview
    The orchestrator manages agent workflow execution as DAGs (Directed Acyclic Graphs).
    It provides endpoints for:
    - **Runs**: Create, start, cancel, and monitor workflow executions
    - **Agents**: Register and manage agent definitions
    - **Flows**: Store and retrieve reusable workflow definitions
    - **Artifacts**: Upload and download run artifacts
    - **Jobs**: Monitor Kubernetes job status (when running in K8s mode)

    ## Event Streaming
    Run events are streamed via Server-Sent Events (SSE) at `/api/v1/runs/{id}/events`.
    Events include logs, checkpoints, node status changes, progress updates, and control flow events.

    ## Authentication
    Currently no authentication is required. In production, requests should be
    proxied through the gateway which handles authentication.
  version: 1.0.0
  contact:
    name: MentatLab Team
  license:
    name: MIT
servers:
  - url: http://localhost:7070
    description: Local development
  - url: http://orchestrator:7070
    description: Kubernetes internal
...
paths:
...
  /api/v1/flows:
    get:
      tags: [flows]
      summary: List flows
      operationId: listFlows
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: created_by
          in: query
          description: Filter by creator
          schema:
            type: string
      responses:
        '200':
          description: List of flows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFlowsResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    post:
      tags: [flows]
      summary: Create a flow
      operationId: createFlow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFlowRequest'
      responses:
        '201':
          description: Flow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Flow already exists
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/flows/{id}:
    get:
      tags: [flows]
      summary: Get flow details
      operationId: getFlow
      parameters:
        - $ref: '#/components/parameters/FlowId'
      responses:
        '200':
          description: Flow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    put:
      tags: [flows]
      summary: Update a flow
      operationId: updateFlow
      parameters:
        - $ref: '#/components/parameters/FlowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFlowRequest'
      responses:
        '200':
          description: Flow updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      tags: [flows]
      summary: Delete a flow
      operationId: deleteFlow
      parameters:
        - $ref: '#/components/parameters/FlowId'
      responses:
        '204':
          description: Flow deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # Artifact management
  /api/v1/runs/{id}/artifacts:
    get:
      tags: [artifacts]
      summary: List run artifacts
      operationId: listRunArtifacts
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: List of artifacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListArtifactsResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    post:
      tags: [artifacts]
      summary: Upload an artifact
      description: |
        Supports two modes:
        1. **Direct upload**: Send `multipart/form-data` with a file field
        2. **Presigned URL**: Send JSON with name/content_type to get an upload URL
      operationId: uploadArtifact
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                node_id:
                  type: string
          application/json:
            schema:
              $ref: '#/components/schemas/UploadArtifactRequest'
      responses:
        '200':
          description: Upload URL generated (JSON mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadURLResponse'
        '201':
          description: Artifact uploaded (multipart mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactRef'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/artifacts:
    get:
      tags: [artifacts]
      summary: Get artifact content
      description: Streams the artifact content directly
      operationId: getArtifact
      parameters:
        - name: uri
          in: query
          required: true
          description: Artifact URI
          schema:
            type: string
      responses:
        '200':
          description: Artifact content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      tags: [artifacts]
      summary: Delete an artifact
      operationId: deleteArtifact
      parameters:
        - name: uri
          in: query
          required: true
          description: Artifact URI
          schema:
            type: string
      responses:
        '204':
          description: Artifact deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/artifacts/download-url:
    post:
      tags: [artifacts]
      summary: Get presigned download URL
      operationId: getArtifactDownloadURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [uri]
              properties:
                uri:
                  type: string
                  description: Artifact URI
      responses:
        '200':
          description: Download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadURLResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # Job management
...
components:
  parameters:
    RunId:
      name: id
      in: path
      required: true
      description: Run ID
      schema:
        type: string
...
    FlowId:
      name: id
      in: path
      required: true
      description: Flow ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: Service or dependency unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable message
        details:
          type: object
          additionalProperties: true

    # Run types
...
    Plan:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeSpec'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeSpec'
    NodeSpec:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          description: Node type (agent, conditional, for_each, subflow)
        agent_id:
          type: string
        command:
          type: array
          items:
            type: string
        image:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
        inputs:
          type: array
          items:
            type: string
          description: IDs of upstream nodes
        timeout:
          type: integer
          description: Timeout in nanoseconds (Go duration)
        retries:
          type: integer
        conditional:
          $ref: '#/components/schemas/ConditionalConfig'
        for_each:
          $ref: '#/components/schemas/ForEachConfig'
        subflow:
          $ref: '#/components/schemas/SubflowConfig'
    EdgeSpec:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
        to:
          type: string
    ConditionalConfig:
      type: object
      properties:
        expression:
          type: string
        then_node:
          type: string
        else_node:
          type: string
    ForEachConfig:
      type: object
      properties:
        collection:
          type: string
        item_var:
          type: string
        body_node:
          type: string
        max_parallel:
          type: integer
    SubflowConfig:
      type: object
      properties:
        flow_id:
          type: string
        inputs:
          type: object
          additionalProperties: true
...
    Flow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        plan:
          $ref: '#/components/schemas/Plan'
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateFlowRequest:
      type: object
      required: [name, plan]
      properties:
        name:
          type: string
        description:
          type: string
        plan:
          $ref: '#/components/schemas/Plan'
    UpdateFlowRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        plan:
          $ref: '#/components/schemas/Plan'
    ListFlowsResponse:
      type: object
      properties:
        flows:
          type: array
          items:
            $ref: '#/components/schemas/Flow'
        count:
          type: integer

    # Artifact types
    ArtifactRef:
      type: object
      properties:
        uri:
          type: string
        size:
          type: integer
          format: int64
        content_type:
          type: string
        created_at:
          type: string
          format: date-time
    ListArtifactsResponse:
      type: object
      properties:
        run_id:
          type: string
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactRef'
        count:
          type: integer
    UploadArtifactRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        content_type:
          type: string
        node_id:
          type: string
    UploadURLResponse:
      type: object
      properties:
        upload_url:
          type: string
        expires_in:
          type: string
        path:
          type: string
    DownloadURLResponse:
      type: object
      properties:
        download_url:
          type: string
        expires_in:
          type: string

    # Job types
...